% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPWR.R
\docType{class}
\name{ModelPWR-class}
\alias{ModelPWR-class}
\alias{ModelPWR}
\title{A Reference Class which represents a fitted PWR model.}
\description{
ModelPWR represents an estimated PWR model.
}
\section{Fields}{

\describe{
\item{\code{paramPWR}}{A \link{ParamPWR} object. It contains the estimated
values of the parameters.}

\item{\code{statPWR}}{A \link{StatPWR} object. It contains all the statistics
associated to the PWR model.}
}}

\section{Methods}{

\describe{
\item{\code{plot(what = c("regressors", "segmentation"))}}{Plot method.
\describe{
  \item{\code{what}}{The type of graph requested:
    \itemize{
      \item \code{"regressors" = } Polynomial regression components.
      \item \code{"segmentation" = } Estimated signal (Estimated
      piecewise regression and transition time points).
    }
  }
}
By default, all the above graphs are produced.}

\item{\code{summary()}}{Summary method.}
}}

\seealso{
\link{ParamPWR}, \link{StatPWR}
}
